LayoutTests/imported/w3c/web-platform-tests/FileAPI/reading-data-section/filereader_readAsDataURL-expected.txt


FAIL FileAPI Test: filereader_readAsDataURL assert_equals: expected 1 but got 2
